Add files via upload
Cards can be a handy way to show content that includes different elements, such as the title, image, buttons, icons, etc. The blocks or cards can have any kind of unique element designed specifically for them based on your content. The key idea here is that adding a card to any site will immediately provide that inventive and creative touch.